bot/requirements.txt
aiogram==2.25.1
psycopg2==2.9.7
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
tabulate==0.8.9